+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ### Fixed
 - Fixed overrides impacting the original dimension not getting picked up for parallel world
+- Fixed Death Recall Token teleporting to the surface or nether roof instead of the actual death location